Implementar filtro year.

+++ app/src/main/java/edu/iesam/bikerly/data/local/mock/Mappers.kt
@@ -8,7 +8,7 @@ fun MotorbikeMockModel.toModel(): Motorbike {
         this.id,
         this.make,
         this.model,
-        this.year,
+        this.year.toInt(),
         this.type,
         this.displacement.toInt(),
         this.img.toUri()
